rename module sphinx_ext
