+++ cylc/sphinx_ext/diff_selection/_static/css/diff_selector.css
