+++ cylc/sphinx_ext/diff_selection/_static/js/diff_selector.js
