+++ cylc/sphinx_ext/hieroglyph_theme_addons/__init__.py
@@ -34,7 +34,7 @@
 """
 
 
-from cylc.sphinx import register_static
+from cylc.sphinx_ext import register_static
 
 
 __version__ = '1.0.0'
