+++ cylc/sphinx_ext/practical/__init__.py
@@ -48,7 +48,7 @@ __all__ = [
 ]
 
 
-from cylc.sphinx.practical.admonitions import (
+from cylc.sphinx_ext.practical.admonitions import (
     Practical,
     PracticalExtension,
     Spoiler
@@ -60,7 +60,7 @@ __version__ = '1.0.0'
 
 def setup(app):
     """Sphinx setup function."""
-    from cylc.sphinx import register_static
+    from cylc.sphinx_ext import register_static
     app.add_directive('practical', Practical)
     app.add_directive('practical-extension', PracticalExtension)
     app.add_directive('spoiler', Spoiler)
