+++ cylc/sphinx_ext/practical/_static/js/spoiler.js
