+++ cylc/sphinx_ext/rtd_theme_addons/_static/js/addons.js
