+++ cylc/sphinx_ext/tests/test_init.py
@@ -1,6 +1,6 @@
 from unittest.mock import Mock
 
-from cylc.sphinx import register_static
+from cylc.sphinx_ext import register_static
 
 
 def test_register_static():
@@ -8,12 +8,12 @@ def test_register_static():
     app.config = Mock()
     app.config.html_static_path = []
 
-    register_static(app, 'cylc.sphinx.foo')
+    register_static(app, 'cylc.sphinx_ext.foo')
     assert app.config.html_static_path[0].endswith(
-        'cylc/sphinx/foo/_static'
+        'cylc/sphinx_ext/foo/_static'
     )
 
-    register_static(app, 'cylc.sphinx.foo.bar')
+    register_static(app, 'cylc.sphinx_ext.foo.bar')
     assert app.config.html_static_path[1].endswith(
-        'cylc/sphinx/foo/bar/_static'
+        'cylc/sphinx_ext/foo/bar/_static'
     )
