+++ pytest.ini
@@ -18,4 +18,4 @@
 addopts = 
     --doctest-modules
 testpaths =
-    cylc/sphinx/
+    cylc/sphinx_ext/
